Fix things so MinGW compiles

+++ libdrizzle/sha1.cc
@@ -128,7 +128,7 @@ SHA1Init(SHA1_CTX *context)
 void
 SHA1Update(SHA1_CTX *context, const uint8_t *data, size_t len)
 {
-	size_t i, j;
+	uint32_t i, j;
 
 	j = (size_t)((context->count >> 3) & 63);
 	context->count += (len << 3);
